requirements.txt
dds-cloudapi-sdk
label-studio-sdk
label-studio-converter
pycocotools
lxml
pillow
numpy